Task commit

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	char c;
 	int i;
 	float f;
-	printf("Size of a char: %lu byte(s).\n", sizeof(i));
-	printf("Size of an int: %lu byte(s).\n", sizeof(d));
+	printf("Size of a char: %lu byte(s).\n", (unsigned long)sizeof(i));
+	printf("Size of an int: %lu byte(s).\n", (unsigned long)sizeof(d));
 	printf("Size of a long int: %lu byte(s).\n", (long)sizeof(i));
 	printf("Size of a long long int: %lu byte(s).\n", (long long)sizeof(i));
-	printf("Size of a float: %lu byte(s).\n", sizeof(f));
+	printf("Size of a float: %lu byte(s).\n", (unsigned long)sizeof(f));
 	return (0);
 }
